move header files to plugins (edit paths commit)
