+++ RecoBTag/DeepFlavour/plugins/DeepFlavourTFJetTagsProducer.cc
@@ -16,7 +16,7 @@
 
 #include "PhysicsTools/TensorFlow/interface/TensorFlow.h"
 
-#include "RecoBTag/DeepFlavour/interface/tensor_fillers.h"
+#include "RecoBTag/DeepFlavour/plugins/tensor_fillers.h"
 
 // Declaration of the data structure that is hold by the edm::GlobalCache.
 // In TensorFlow, the computational graph is stored in a stateless meta graph object which can be
