+++ RecoBTag/DeepFlavour/src/sorting_modules.cc
@@ -1,5 +1,5 @@
 
-#include "RecoBTag/DeepFlavour/interface/sorting_modules.h"
+#include "RecoBTag/DeepFlavour/plugins/sorting_modules.h"
 #include "FWCore/Utilities/interface/Exception.h"
 #include <iostream>
